+++ back/src/main/java/com/group4/aldalka/domain/post/entity/Ingredient.java
@@ -1,4 +1,4 @@
-package com.group4.aldalka.domain.post.service;
+package com.group4.aldalka.domain.post.entity;
 
 import com.group4.aldalka.domain.BaseEntity;
 import jakarta.persistence.*;
@@ -9,20 +9,20 @@ import java.util.List;
 
 @Getter
 @Entity
-@Table(name = "base_liquior")
+@Table(name = "ingredient")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class BaseLiquor extends BaseEntity {
+public class Ingredient extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "base_liquor_id")
-    private Long baseLiquorId;
+    @Column(name = "ingredient_id")
+    private Long ingredientId;
 
     @Column(name = "name")
     private String name;
 
-    @OneToMany(mappedBy = "baseLiquor", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<PostBaseLiquor> postBaseLiquors = new ArrayList<>();
+    @OneToMany(mappedBy = "ingredient", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<PostIngredient> postIngredients = new ArrayList<>();
 }
