+++ back/src/main/java/com/group4/aldalka/domain/post/entity/Post.java
@@ -1,4 +1,4 @@
-package com.group4.aldalka.domain.post.service;
+package com.group4.aldalka.domain.post.entity;
 
 import com.group4.aldalka.domain.BaseEntity;
 import com.group4.aldalka.domain.user.User;
