+++ back/src/main/java/com/group4/aldalka/domain/post/entity/PostBaseLiquor.java
@@ -1,4 +1,4 @@
-package com.group4.aldalka.domain.post.service;
+package com.group4.aldalka.domain.post.entity;
 
 import com.group4.aldalka.domain.BaseEntity;
 import jakarta.persistence.*;
@@ -6,22 +6,23 @@ import lombok.*;
 
 @Getter
 @Entity
-@Table(name = "post_ingredient")
+@Table(name = "post_base_liquor")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class PostIngredient extends BaseEntity {
+public class PostBaseLiquor extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_ingredient_id")
-    private Long postIngredientId;
+    @Column(name = "post_base_liquor_id")
+    private Long postBaseLiquorId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "ingredient_id", nullable = false)
-    private Ingredient ingredient;
+    @JoinColumn(name = "base_liquor_id", nullable = false)
+    private BaseLiquor baseLiquor;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id", nullable = false)
     private Post post;
 }
+
