+++ back/src/main/java/com/group4/aldalka/domain/post/entity/PostIngredient.java
@@ -1,4 +1,4 @@
-package com.group4.aldalka.domain.post.service;
+package com.group4.aldalka.domain.post.entity;
 
 import com.group4.aldalka.domain.BaseEntity;
 import jakarta.persistence.*;
@@ -6,23 +6,22 @@ import lombok.*;
 
 @Getter
 @Entity
-@Table(name = "post_base_liquor")
+@Table(name = "post_ingredient")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class PostBaseLiquor extends BaseEntity {
+public class PostIngredient extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_base_liquor_id")
-    private Long postBaseLiquorId;
+    @Column(name = "post_ingredient_id")
+    private Long postIngredientId;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "base_liquor_id", nullable = false)
-    private BaseLiquor baseLiquor;
+    @JoinColumn(name = "ingredient_id", nullable = false)
+    private Ingredient ingredient;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id", nullable = false)
     private Post post;
 }
-
